Add test and check for executaltion, mode eval_function
