+++ swarmcg/simulations/__init__.py
@@ -1,2 +1,2 @@
 from .simulation_steps import select_class
-from .runner import generate_steps, ns_to_runner
+from .runner import generate_steps, ns_to_runner, SimulationStep
